Refactor: add CreateStatementInfo.
