+++ server/src/services/definition.ts
@@ -1,11 +1,15 @@
 import {
   DefinitionLink,
+  LocationLink,
   Logger,
   Position,
+  URI,
 } from "vscode-languageserver"
 import { TextDocument } from "vscode-languageserver-textdocument"
 
-import { DefinitionsManager } from "@/server/definitionsManager"
+import { parseCreateStatements } from "@/postgres/parsers/parseCreateStatements"
+import { parseStmtements } from "@/postgres/parsers/statement"
+import { DefinitionCandidate, DefinitionsManager } from "@/server/definitionsManager"
 import { sanitizeWordCandidates } from "@/utilities/sanitizeWord"
 import { getWordRangeAtPosition } from "@/utilities/text"
 
@@ -34,4 +38,56 @@ export async function getDefinitionLinks(
   }
 
   return undefined
+}
+
+
+export async function parseDefinitions(
+  fileText: string,
+  uri: URI,
+  defaultSchema: string,
+): Promise< DefinitionCandidate[] | undefined> {
+  const statements = await parseStmtements(fileText)
+  if (statements === undefined) {
+    return undefined
+  }
+
+
+  return parseCreateStatements(fileText, statements).flatMap(
+    (statementInfo) => {
+      return makeMultiSchemaDefinitionCandidates(
+        statementInfo.name,
+        LocationLink.create(
+          uri,
+          statementInfo.targetRange,
+          statementInfo.targetSelectionRange,
+        ),
+        statementInfo.schema,
+        defaultSchema,
+      )
+    },
+  )
+}
+
+function makeMultiSchemaDefinitionCandidates(
+  definitionName: string,
+  definitionLink: DefinitionLink,
+  schema: string | undefined,
+  defaultSchema: string,
+): DefinitionCandidate[] {
+  const candidates = [
+    {
+      definition: (schema || defaultSchema) + "." + definitionName,
+      definitionLink,
+    },
+  ]
+
+  // On the default schema, add candidate without schema.
+  if (schema === undefined || schema === defaultSchema) {
+    candidates.push({
+      definition: definitionName,
+      definitionLink,
+    })
+  }
+
+  return candidates
 }
